@@ -1,4 +1,5 @@
+/* eslint-disable @typescript-eslint/no-explicit-any */
 export interface Config {
   [key: string]: any;
-  sqs?: {[key: string]: any } ;
+  sqs?: { [key: string]: any };
 }
