@@ -1,9 +1,12 @@
-import { PromiseResult } from 'aws-sdk/lib/request';
-import { ReceiveMessageResult, SendMessageResult } from 'aws-sdk/clients/sqs';
+/* eslint-disable @typescript-eslint/no-unsafe-return */
+/* eslint-disable @typescript-eslint/no-unsafe-argument */
+/* eslint-disable @typescript-eslint/no-explicit-any */
 import { AWSError } from 'aws-sdk';
+import { ReceiveMessageResult, SendMessageResult } from 'aws-sdk/clients/sqs';
+import { PromiseResult } from 'aws-sdk/lib/request';
+import { SQService } from '../../src/services/SQService';
 import { StreamService } from '../../src/services/StreamService';
 import { SQMockClient } from '../models/SQMockClient';
-import { SQService } from '../../src/services/SQService';
 import event from '../resources/stream-event.json';
 
 describe('atf-gen-init', () => {
@@ -45,7 +48,7 @@ describe('atf-gen-init', () => {
     context('when adding a record to the queue', () => {
       it('should successfully add the records to the queue', () => {
         const sendMessagePromises: Array<
-        Promise<PromiseResult<SendMessageResult, AWSError>>
+          Promise<PromiseResult<SendMessageResult, AWSError>>
         > = [];
 
         // eslint-disable-next-line @typescript-eslint/no-unsafe-call,@typescript-eslint/no-unsafe-member-access
@@ -57,15 +60,17 @@ describe('atf-gen-init', () => {
 
         expect.assertions(0);
         return Promise.all(sendMessagePromises).catch((error: AWSError) => {
+          // eslint-disable-next-line jest/no-conditional-expect
           expect(error).toBeFalsy();
         });
       });
 
-      it('should successfully read the added records from the queue', () => sqService
-        .getMessages()
-        .then((messages: ReceiveMessageResult) => {
+      it('should successfully read the added records from the queue', () =>
+        sqService.getMessages().then((messages: ReceiveMessageResult) => {
           // eslint-disable-next-line @typescript-eslint/no-unsafe-return
-          const actual = messages.Messages!.map((message) => JSON.parse(message.Body || ""));
+          const actual = messages.Messages!.map((message) =>
+            JSON.parse(message.Body || ''),
+          );
           expect(actual).toEqual(processedEvent);
         }));
     });
